+++ study_gmm/gmm.py
@@ -12,7 +12,7 @@ from tqdm import tqdm
 from tqdm.contrib.logging import logging_redirect_tqdm
 import logging
 
-
+logger = logging.getLogger(__name__)
 
 class GaussianMixtureModel():
     """Definition of Gaussian Mixture Model (GMM)
@@ -189,15 +189,7 @@ def train_gmm(gmm:GaussianMixtureModel, X:np.ndarray, max_it:int =12):
     return gmm, loglikelihood_history
 
 
-def plot_loglikelihood_history(loglikelihood_history):
-    fig, ax = plt.subplots(1, 1, figsize=(10, 4))
-    ax.plot(range(0, len(loglikelihood_history)), loglikelihood_history, color='k', linestyle='-', marker='o')
-    ax.set_xlim([0,len(loglikelihood_history)])
-    ax.set_ylabel("log liklihood")
-    ax.set_xlabel("iteration step")
-    #plt.ylim([40, 80])
-    ax.grid(True)
-    return fig
+
 
 def load_from_file(input_file:str):
     with open(input_file, 'rb') as f:
